fix error on CampaignForm (when campaign prop not provided) + minor refactors
