+++ frontend/src/atoms/Select.tsx
@@ -1,2 +1,6 @@
 import ReactSelect from "react-select";
 export const Select = (props) => <ReactSelect {...props} />;
+export type Option = {
+  value: string;
+  label: string;
+};
